pyproject.toml
[project]
name = "kg-extractor"
version = "0.1.0"
description = "Extract text from images using NVIDIA API for knowledge graph construction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "python-docx>=1.1.0",
    "PyPDF2>=3.0.0",
    "python-pptx>=0.6.21",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
    "pymupdf>=1.23.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "groq>=1.0.0",
    "qdrant-client>=1.12.0",
    "neo4j>=5.0.0",
]

[project.scripts]
kg-extractor = "kg_extractor.main:main"
kg-agent = "kg_extractor.agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
